Move EncodingPolyfills.cs to Common/src/Polyfills/ directory
Agent-Logs-Url: https://github.com/dotnet/runtime/sessions/06841447-52ae-4051-badf-068f74c098a8
