+++ src/libraries/Microsoft.Extensions.Logging.Console/src/Microsoft.Extensions.Logging.Console.csproj
@@ -40,7 +40,7 @@
     <Compile Include="$(CoreLibSharedDir)System\Diagnostics\CodeAnalysis\UnconditionalSuppressMessageAttribute.cs" />
     <Compile Include="$(CoreLibSharedDir)System\Runtime\CompilerServices\CompilerLoweringPreserveAttribute.cs" />
     <Compile Include="$(CommonPath)Polyfills\TextWriterPolyfills.cs" Link="Common\Polyfills\TextWriterPolyfills.cs" />
-    <Compile Include="$(CommonPath)System\Text\EncodingPolyfills.cs" Link="Common\Polyfills\EncodingPolyfills.cs" />
+    <Compile Include="$(CommonPath)Polyfills\EncodingPolyfills.cs" Link="Common\Polyfills\EncodingPolyfills.cs" />
   </ItemGroup>
 
   <ItemGroup>
